requirements.txt
pyqt5
pyserial